Move function to convert table structure substance to raw ephemeral form into add-edit-modal
