+++ mathesar_ui/src/systems/data-forms/form-maker/data-form-utilities/transformers.ts
@@ -2,8 +2,6 @@ import {
   type RawDataForm,
   type RawDataFormField,
   type RawDataFormSource,
-  type RawEphemeralDataForm,
-  dataFormStructureVersion,
 } from '@mathesar/api/rpc/forms';
 import type { TableStructureSubstance } from '@mathesar/stores/table-data/TableStructure';
 import { getGloballyUniqueId } from '@mathesar-component-library';
@@ -87,7 +85,7 @@ function getColumnDetailFromFormSource(
   return columnInfo;
 }
 
-export function rawEphemeralFieldToEphemeralFieldProps(
+function rawEphemeralFieldToEphemeralFieldProps(
   rawEphemeralField: RawDataFormField,
   parentTableOid: number,
   formSource: RawDataFormSource,
@@ -166,76 +164,5 @@ export function rawDataFormToDataFormStructureProps(
         formSource,
       ),
     ),
-  };
-}
-
-function fieldColumnToRawDataFormField(
-  fc: FieldColumn,
-  tableStructureSubstance: TableStructureSubstance,
-  index: number,
-): RawDataFormField {
-  const baseProps = {
-    key: getGloballyUniqueId(),
-    label: fc.column.name,
-    help: null,
-    index,
-    is_required: false,
-    column_attnum: fc.column.id,
-    styling: {},
-  };
-  if (fc.foreignKeyLink) {
-    const referentTableOid = fc.foreignKeyLink.relatedTableOid;
-    const referenceTableName = tableStructureSubstance.linksInTable.find(
-      (lnk) => lnk.table.oid === referentTableOid,
-    )?.table.name;
-    return {
-      ...baseProps,
-      kind: 'foreign_key',
-      label: referenceTableName ?? baseProps.label,
-      related_table_oid: referentTableOid,
-      fk_interaction_rule: 'must_pick',
-      child_fields: [],
-    };
-  }
-  return {
-    ...baseProps,
-    kind: 'scalar_column',
-  };
-}
-
-function tableStructureSubstanceToRawEphemeralField(
-  tableStructureSubstance: TableStructureSubstance,
-): RawDataFormField[] {
-  return [...tableStructureSubstance.processedColumns.values()]
-    .filter((pc) => !pc.column.default?.is_dynamic)
-    .map((c, index) => {
-      const ef = fieldColumnToRawDataFormField(
-        FieldColumn.fromProcessedColumn(c),
-        tableStructureSubstance,
-        index,
-      );
-      return ef;
-    });
-}
-
-export function tableStructureSubstanceRawEphemeralForm(
-  tableStructureSubstance: TableStructureSubstance,
-): RawEphemeralDataForm {
-  return {
-    base_table_oid: tableStructureSubstance.table.oid,
-    schema_oid: tableStructureSubstance.table.schema.oid,
-    database_id: tableStructureSubstance.table.schema.database.id,
-    version: dataFormStructureVersion,
-    name: tableStructureSubstance.table.name,
-    description: null,
-    header_title: {
-      text: tableStructureSubstance.table.name,
-    },
-    header_subtitle: null,
-    associated_role_id: null,
-    submit_message: null,
-    submit_redirect_url: null,
-    submit_button_label: null,
-    fields: tableStructureSubstanceToRawEphemeralField(tableStructureSubstance),
   };
 }
